Move ConnectWallet to separate module
